@@ -6,18 +6,18 @@ import 'package:flutter/services.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
+  const MyApp({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flachwitze',
-      theme: ThemeData(
-        primaryColor: Colors.deepOrange,
-        accentColor: Colors.deepOrange,
+      theme: ThemeData.from(
+        colorScheme: ColorScheme.fromSwatch(primarySwatch: Colors.deepOrange),
       ),
       home: MainPage(),
     );
@@ -25,7 +25,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MainPage extends StatefulWidget {
-  MainPage({Key key}) : super(key: key);
+  MainPage({Key? key}) : super(key: key);
 
   final Random random = Random();
 
@@ -34,10 +34,10 @@ class MainPage extends StatefulWidget {
 }
 
 class _MainPageState extends State<MainPage> {
-  List<String> _jokes;
+  late List<String> _jokes;
   int _idx = -1;
   String _title = 'Scherzfragen';
-  String _joke;
+  String? _joke;
 
   @override
   void initState() {
@@ -64,77 +64,80 @@ class _MainPageState extends State<MainPage> {
       ),
       backgroundColor: Colors.orange,
       drawer: Drawer(
-        child: Column(
-          children: [
-            Expanded(
-              child: ListView(
-                padding: EdgeInsets.zero,
-                children: <Widget>[
-                  ListTile(
-                    title: Text('Scherzfragen'),
-                    onTap: () {
-                      loadJokes();
-                      Navigator.pop(context);
-                    },
-                  ),
-                  ListTile(
-                    title: Text('Mutterwitze'),
-                    onTap: () {
-                      loadJokes(name: 'deine-mutter');
-                      Navigator.pop(context);
-                    },
-                  ),
-                  ListTile(
-                    title: Text('Chuck Norris Witze'),
-                    onTap: () {
-                      loadJokes(name: 'chuck-norris');
-                      Navigator.pop(context);
-                    },
-                  ),
-                  ListTile(
-                    title: Text('Sprüche'),
-                    onTap: () {
-                      loadJokes(name: 'sprueche');
-                      Navigator.pop(context);
-                    },
-                  ),
-                  ListTile(
-                    title: Text('Gespräche'),
-                    onTap: () {
-                      loadJokes(name: 'gespraeche');
-                      Navigator.pop(context);
-                    },
-                  ),
-                  ListTile(
-                    title: Text('Namenswitze'),
-                    onTap: () {
-                      loadJokes(name: 'namen');
-                      Navigator.pop(context);
-                    },
-                  ),
-                  ListTile(
-                    title: Text('Anti-Witze'),
-                    onTap: () {
-                      loadJokes(name: 'anti');
-                      Navigator.pop(context);
-                    },
-                  ),
-                ],
+        child: Container(
+          color: Colors.white,
+          child: Column(
+            children: [
+              Expanded(
+                child: ListView(
+                  padding: EdgeInsets.zero,
+                  children: <Widget>[
+                    ListTile(
+                      title: const Text('Scherzfragen'),
+                      onTap: () {
+                        loadJokes();
+                        Navigator.pop(context);
+                      },
+                    ),
+                    ListTile(
+                      title: const Text('Mutterwitze'),
+                      onTap: () {
+                        loadJokes(name: 'deine-mutter');
+                        Navigator.pop(context);
+                      },
+                    ),
+                    ListTile(
+                      title: const Text('Chuck Norris Witze'),
+                      onTap: () {
+                        loadJokes(name: 'chuck-norris');
+                        Navigator.pop(context);
+                      },
+                    ),
+                    ListTile(
+                      title: const Text('Sprüche'),
+                      onTap: () {
+                        loadJokes(name: 'sprueche');
+                        Navigator.pop(context);
+                      },
+                    ),
+                    ListTile(
+                      title: const Text('Gespräche'),
+                      onTap: () {
+                        loadJokes(name: 'gespraeche');
+                        Navigator.pop(context);
+                      },
+                    ),
+                    ListTile(
+                      title: const Text('Namenswitze'),
+                      onTap: () {
+                        loadJokes(name: 'namen');
+                        Navigator.pop(context);
+                      },
+                    ),
+                    ListTile(
+                      title: const Text('Anti-Witze'),
+                      onTap: () {
+                        loadJokes(name: 'anti');
+                        Navigator.pop(context);
+                      },
+                    ),
+                  ],
+                ),
               ),
-            ),
-            _madeWithFlutter(),
-            _hostedOnGithub(),
-          ],
+              _madeWithFlutter(),
+              _hostedOnGithub(),
+            ],
+          ),
         ),
       ),
       body: Padding(
         padding: const EdgeInsets.all(16.0),
         child: Center(
           child: _joke == null
-              ? CircularProgressIndicator()
+              ? const CircularProgressIndicator()
               : SelectableText(
-                  _joke,
-                  style: TextStyle(fontSize: 48.0, color: Colors.white),
+                  _joke!,
+                  style: const TextStyle(fontSize: 48.0, color: Colors.white),
                   textAlign: TextAlign.center,
                 ),
         ),
@@ -145,13 +148,13 @@ class _MainPageState extends State<MainPage> {
           FloatingActionButton(
             onPressed: _randomJoke,
             tooltip: 'Random',
-            child: Icon(Icons.refresh),
+            child: const Icon(Icons.refresh),
           ),
-          SizedBox(width: 16),
+          const SizedBox(width: 16),
           FloatingActionButton(
             onPressed: _next,
             tooltip: 'Next',
-            child: Icon(Icons.arrow_forward),
+            child: const Icon(Icons.arrow_forward),
           ),
         ],
       ),
@@ -164,21 +167,21 @@ class _MainPageState extends State<MainPage> {
       child: RichText(
         text: TextSpan(
           children: [
-            TextSpan(text: 'Made with '),
+            const TextSpan(text: 'Made with '),
             TextSpan(
               children: [
                 TextSpan(
                   text: 'Flutter',
-                  style: new TextStyle(color: Colors.blue),
+                  style: const TextStyle(color: Colors.blue),
                   mouseCursor: SystemMouseCursors.click,
                   recognizer: TapGestureRecognizer()
                     ..onTap = () {
                       launch('https://flutter.dev');
                     },
                 ),
-                WidgetSpan(
+                const WidgetSpan(
                   child: Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 2.0),
+                    padding: EdgeInsets.symmetric(horizontal: 2.0),
                     child: Icon(
                       Icons.open_in_new,
                       size: 16.0,
@@ -223,21 +226,21 @@ class _MainPageState extends State<MainPage> {
       child: RichText(
         text: TextSpan(
           children: [
-            TextSpan(text: 'Hosted on '),
+            const TextSpan(text: 'Hosted on '),
             TextSpan(
               children: [
                 TextSpan(
                   text: 'Github',
-                  style: new TextStyle(color: Colors.blue),
+                  style: const TextStyle(color: Colors.blue),
                   mouseCursor: SystemMouseCursors.click,
                   recognizer: TapGestureRecognizer()
                     ..onTap = () {
                       launch('https://github.com/mfietz/flachwitze');
                     },
                 ),
-                WidgetSpan(
+                const WidgetSpan(
                   child: Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 2.0),
+                    padding: EdgeInsets.symmetric(horizontal: 2.0),
                     child: Icon(
                       Icons.open_in_new,
                       size: 16.0,
@@ -255,15 +258,11 @@ class _MainPageState extends State<MainPage> {
 
 extension CapitalizedStringExtension on String {
   String toTitleCase() {
-    if (this == null) {
-      return null;
+    if (length <= 1) {
+      return toUpperCase();
     }
 
-    if (this.length <= 1) {
-      return this.toUpperCase();
-    }
-
-    final List<String> words = this.split(' ');
+    final List<String> words = split(' ');
 
     final capitalizedWords = words.map((word) {
       if (word.trim().isNotEmpty) {
